refactor: 상수, 유틸 파일 분리 및 jsx 확장자 변경
