--- src/constants.js
@@ -1,5 +1,0 @@
-export const OPERAND_ERROR_VALUE = '오류';
-
-export const OPERAND_MAX_LENGTH = 3;
-
-export const STORAGE_KEY = 'state';
