+++ src/constants/index.js
@@ -1,0 +1,8 @@
+const STORAGE_KEY = 'state';
+
+const OPERAND_ERROR_VALUE = '오류';
+const OPERAND_MAX_LENGTH = 3;
+
+const DIGITS = [9, 8, 7, 6, 5, 4, 3, 2, 1, 0];
+
+export { STORAGE_KEY, OPERAND_ERROR_VALUE, OPERAND_MAX_LENGTH, DIGITS };
