+++ src/util/index.js
@@ -1,0 +1,8 @@
+const operation = {
+  '+': (firstNumber, secondNumber) => firstNumber + secondNumber,
+  '-': (firstNumber, secondNumber) => firstNumber - secondNumber,
+  'x': (firstNumber, secondNumber) => firstNumber * secondNumber,
+  '/': (firstNumber, secondNumber) => Math.floor(firstNumber / secondNumber),
+};
+
+export { operation };
